Fixed accessor to mutation rate to return double.
git-svn-id: http://wlab.ccbb.utexas.edu/svn/evoli@94 dadfaeae-b11a-0410-95f2-80a8e7ca9aa5

+++ trunk/src/evolver/population.hh
@@ -203,7 +203,7 @@ public:
 	 * Retrieves the mutation rate.
 	 * @return The mutation rate.
 	 **/
-	int getMutationRate() { return m_U; }
+	double getMutationRate() { return m_U; }
 
 	/**
 	 * Advances the population one time step.
